Todos CRUD feito, mas somente Jogador no fluxo de service

+++ projeto-campeonato-volei/src/main/java/com/example/projeto/campeonato/volei/repository/TimeRepository.java
@@ -3,7 +3,7 @@ package com.example.projeto.campeonato.volei.repository;
 import com.example.projeto.campeonato.volei.domain.Time;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimeRepository extends JpaRepository<Time, Long> {
+public interface TimeRepository extends JpaRepository<Time, Integer> {
 
     boolean existsById(Integer id);
